npm run test works w/ new refactor
